refactor: voice banner text hook 분리
